Registration and Login method (backend). Now works properly.

+++ src/main/java/com/example/iti0302backend/service/PostService.java
@@ -2,7 +2,6 @@ package com.example.iti0302backend.service;
 
 import com.example.iti0302backend.post.Post;
 import com.example.iti0302backend.post.PostRepository;
-import com.example.iti0302backend.user.User;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -16,7 +15,7 @@ public class PostService {
         this.postRepository = postRepository;
     }
 
-    public Optional<Post> findById(Integer id){
+    public Optional<Post> findById(Long id){
         return postRepository.findById(id);
     }
